Rename utils.py to invocation_helpers.py per review feedback.
