+++ src/rai_core/rai/agents/langchain/core/conversational_agent.py
@@ -29,7 +29,7 @@ from langgraph.graph.state import CompiledStateGraph
 from langgraph.prebuilt.tool_node import tools_condition
 
 from rai.agents.langchain.core.tool_runner import ToolRunner
-from rai.agents.langchain.utils import invoke_llm_with_tracing
+from rai.agents.langchain.invocation_helpers import invoke_llm_with_tracing
 
 
 class State(TypedDict):
