+++ src/rai_core/rai/agents/langchain/invocation_helpers.py
